Cleanup tests
This commit does a couple of things. The first is that it fixes an intermittent failure with one of the tests. It was failing because the subsequent tests added a second circuit_method call to the `second_method` function, which caused it to count two failures for each error. So, depending on the order in which the tests were run, it could fail. The other thing it does is update the rspec syntax to use the `expect` style so that the deprecation warnings go away.