Allow percentage failures
This breaks out the logic that determines if the circuit breaker is tripped into another class that responds to the `tripped?(circuit_state)` method. It then adds a percentage-based trip checker that will flip when a certain percentage of the given calls fail as opposed to a hard-coded number of failures.